requirements.txt
asgiref==3.6.0
branca==0.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
decorator==5.1.1
dj-database-url==1.3.0
Django==4.2
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
folium==0.14.0
future==0.18.3
geocoder==1.38.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.6.0
pytz==2023.3
ratelim==0.1.6
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
